requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-marshmallow==1.0.0
marshmallow-sqlalchemy==1.0.0
